requirements.txt
Django==5.2.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-filter==24.3
django-cors-headers==4.3.1
python-dotenv==1.0.0
dj-database-url==2.1.0
Pillow==10.4.0
psycopg==3.2.1
psycopg-binary==3.2.1